feat:  buttonLabel && encapsulate content
